+++ README.md
@@ -1,0 +1,17 @@
+# curve-game
+
+> Proof of concept game like that: http://curvefever.com/
+
+## Install
+
+If you don't have global package:
+
+```
+npm install webpack
+```
+
+Run command to build `dist/bundle.js`
+
+```
+webpack
+```
