+++ app/index.html
@@ -1,0 +1,16 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <title>curve-game</title>
+    <link rel="stylesheet" href="styles/main.css">
+</head>
+<body>
+    <div class="game">
+        <canvas class="playground"></canvas>
+        <div class="info hidden">Game over</div>
+    </div>
+
+    <script src="dist/bundle.js"></script>
+</body>
+</html>
