+++ app/scripts/models/Bonus.js
@@ -1,0 +1,4 @@
+function Bonus() {
+}
+
+module.exports = Bonus;
