--- css/style.css
@@ -1,28 +1,0 @@
-.hidden { display: none; }
-
-div.game {
-	position: relative;
-	width: 500px;
-	height: 500px;
-}
-
-div.game canvas.playground {
-	width: inherit;
-	height: inherit;
-	border: 3px solid #000;
-}
-
-div.game div.info {
-	width: 200px;
-	height: 100px;
-	line-height: 100px;
-	text-align: center;
-	position: absolute;
-	top: 50%;
-	left: 50%;
-	margin-left: -100px;
-	margin-top: -50px;
-	border: 3px solid #000;
-	background-color: #FFF;
-	font-size: 24px;
-}
