--- index.html
@@ -1,21 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-	<meta charset="UTF-8">
-
-	<title>curve-game</title>
-
-	<link rel="stylesheet" href="css/style.css">
-</head>
-<body>
-	<div class="game">
-		<canvas class="playground"></canvas>
-		<div class="info hidden">Game over</div>
-	</div>
-
-	<script src="src/Game.js"></script>
-	<script src="src/Player.js"></script>
-	<script src="src/Bonus.js"></script>
-	<script src="src/run.js"></script>
-</body>
-</html>
