--- src/Bonus.js
@@ -1,1 +1,0 @@
-function Bonus() {}
