+++ webpack.config.js
@@ -1,0 +1,8 @@
+module.exports = {
+    entry: './app/scripts/main',
+
+    output: {
+        filename: 'bundle.js',
+        path: './app/dist/'
+    }
+};
